tf_serving_perf_test_classifier/client/requirements.txt
requests
aiohttp
asyncio
Pillow
keras==2.3.1
tensorflow==2.2.0